Added functionality(expand for details), starting to work on statistic part soon(hopefully)
*quickly written, potential typos* Functionality is:

Whenever the user searches for a serie, it will add a local JSON file with the result of the search inside. Most commonly there are 10 results, so there will be 10 local json files created with those series inside. This will create a bunch of files and they will be used later. I am aware that if a serie is ongoing, that local JSON file is incomplete(since the serie isnt completed/finished yet), and so far it isnt connected to the API. So I made another file called ToCheck.json, that file contains all the 'unfinished' serie names. Then i made a function that will cycle through those unfinished series once per day and update them if they finish(so far it will only update once the serie is finished, i actually realised that it was unproductive while reading this, i will most definetly make it so that it updates the series if there is new data on it, not just wait until it finishes and then update it)

All of the local serie JSON files i mentioned earlier as well as the database are now combined to autofill results for the textbox. As well as a huge improvement. You can now SHIFT+ENTER if you have the autofill option there(and if the serie is saved in a local json file), this will prevent the C# application from making the API call and use the local file instead. This saves KITSU ever so slightly of requests, and improves the speed and feel of the program drastically. Since the api request takes about 1-2 seconds, if you now have the file saved locally, it will just use that and not make the API call, so it feels practically instant. I am very happy about this improvement.

I will also fix the issue that is currently open in the near future, the one regarding default database selection, since i dont think that is optimal at the moment. I do wish to complete this before going on to the statistics part. Also i currently have no idea how i am going to do the statistics and display them to the user, in the start it might just be text and labels since i currently have no way of doing it in a visually appealing way. I could search it up since im pretty sure such things exist, but i would like to come up with a cool solution myself, since i've managed to create all this without google so far(with the exception of errors and stuff general google stuff that devs cant live without, i dont google the consept of my program or how to do statistics in c#, all the functions are self made so far and i would wish to keep it that way*well i must point out that newtonsoft's library and mysql.data.dll is obviously not mine but you get the point)