Added a new form that will display the last episode in every serie you watched.
So far it sorts by alphabet and you can search/filter by typing in the textbox. I want it to sort by date, and i'll see if i can make that happen, i havent put any thought into it yet. I'm also not sure if it should sort alphabetically when filtering and if this should be customizable. Could also sort out series that are finished and that the user has watched to the end to the bottom, maybe let the user decide?