I am close to satisfied with LatestWatched's functionality
I am considering to implement more things when it comes to getting all episodes between 2 dates or 'last weekend' etc. However the problem lies with how the user shall pick this. I can code in the functionality but for it not to get too clunky or just filling the form with controls i am holding it back a bit till i find a satisfactory approach